init: chat feat: fix: logic for random button
