+++ helpers/action_functions/enemyAttck.js
@@ -1,5 +1,6 @@
 import {getIIPoints} from "../helth_points/getIIPoints.js";
 import {getRandomNumber} from "../getRandomNumber.js";
+import {HIT} from "../../store/hit_points.js";
 
 
 export function enemyAttack() {
